feat: implement job management API with CRUD operations
- Added jobController.js to handle job-related logic including:
  - getAllJobs: retrieves all jobs
  - createJob: creates a new job
  - getJob: retrieves a specific job by ID
  - updateJob: updates an existing job
  - deleteJob: deletes a job by ID

- Added jobRouter.js to define routes for job management:
  - GET /jobs: fetch all jobs
  - POST /jobs: create a new job
  - GET /jobs/:id: fetch a job by ID
  - PATCH /jobs/:id: update a job by ID
  - DELETE /jobs/:id: delete a job by ID
