+++ controllers/jobController.js
@@ -1,0 +1,62 @@
+import { nanoid } from "nanoid";
+let jobs = [
+  { id: nanoid(), company: "Apple", position: "front-end" },
+  { id: nanoid(), company: "google", position: "backend" },
+];
+
+export const getAllJobs = async (req, res) => {
+  res.status(200).json({ jobs });
+};
+
+export const createJob = async (req, res) => {
+  const { company, position } = req.body;
+  if (!company || !position) {
+    return res.status(400).json({ msg: "please provide company & position" });
+  }
+  const id = nanoid(10);
+  const job = { id, company, position };
+  jobs.push(job);
+  res.status(201).json({ job });
+};
+
+export const getJob = async (req, res) => {
+  const { id } = req.params;
+  const job = jobs.find((job) => job.id === id);
+  if (!job) {
+    return res
+      .status(400)
+      .json({ msg: `the job for the id:${id} is not found` });
+  }
+  res.status(200).json({ job });
+};
+
+export const updateJob = async (req, res) => {
+  const { company, position } = req.body;
+  if (!company || !position) {
+    return res.status(400).json({ msg: "please provide company & position" });
+  }
+  const { id } = req.params;
+  const job = jobs.find((job) => job.id === id);
+  if (!job) {
+    return res
+      .status(400)
+      .json({ msg: `the job for the id:${id} is not found` });
+  }
+  job.company = company;
+  job.position = position;
+  res.status(200).json({ job });
+};
+
+export const deleteJob = async (req, res) => {
+  const { id } = req.params;
+  const job = jobs.find((job) => job.id === id);
+  if (!job) {
+    return res
+      .status(400)
+      .json({ msg: `the job for the id:${id} is not found` });
+  }
+
+  const newJobs = jobs.filter((job) => job.id !== id);
+  jobs = newJobs;
+  res.status(200).json({ msg: "Job Deleted" });
+};
